fix registerRestRoute types (#10118)
* fix registerRestRoute types

* Make a RestRouteDefinition available in vates/types

---------
