+++ @vates/types/package.json
@@ -32,7 +32,8 @@
   "devDependencies": {
     "@eslint/js": "^9.19.0",
     "typescript": "~5.6.3",
-    "typescript-eslint": "^8.61.0"
+    "typescript-eslint": "^8.61.0",
+    "@types/express": "^5.0.0"
   },
   "scripts": {
     "build": "tsc",
