+++ @vates/types/src/index.mts
@@ -5,3 +5,4 @@ export * from './xo-app.mjs'
 export * from './lib/xen-orchestra-xapi.mjs'
 export * from './lib/xen-orchestra-acl.mjs'
 export * from './lib/complex-matcher.mjs'
+export * from './lib/rest-api.mjs'
