+++ @xen-orchestra/rest-api/src/router/types.mts
@@ -1,9 +1,11 @@
-import type { SecurityName } from '../middlewares/authentication.middleware.mjs'
 import type { AclEntry } from '../middlewares/acl.middleware.mjs'
-import type { NextFunction, Request, Response } from 'express'
 import type { MaybePromise } from '../helpers/helper.type.mjs'
 import type { RestApi } from '../rest-api/rest-api.mjs'
 import type { VatesTask } from '@vates/types/lib/vates/task'
+import type { BaseRouteDefinition, FieldDefinition, ParamFieldDefinition, QueryFieldDefinition } from '@vates/types'
+import type { NextFunction, Request, Response } from 'express'
+
+export type { FieldDefinition, ParamFieldDefinition, QueryFieldDefinition }
 
 export type CreateAction = <CbType>(
   cb: (task: VatesTask) => MaybePromise<CbType>,
@@ -26,62 +28,7 @@ export type MiddlewareDescriptor =
   | { name: 'json' | 'urlencoded' | 'text' | 'raw'; options?: Record<string, unknown> }
   | { name: 'acl'; acls: AclEntry | AclEntry[] }
 
-export type FieldDefinition =
-  | {
-      type: 'string'
-      example?: string
-      optional?: boolean
-    }
-  | {
-      type: 'boolean'
-      example?: boolean
-      optional?: boolean
-    }
-  | {
-      type: 'number'
-      example?: number
-      optional?: boolean
-    }
-  | {
-      type: 'enum'
-      enum: string[]
-      example?: string
-      optional?: boolean
-    }
-  | {
-      type: 'object'
-      fields: Record<string, FieldDefinition>
-      optional?: boolean
-    }
-  | {
-      type: 'array'
-      items: FieldDefinition
-      example?: unknown[]
-      optional?: boolean
-    }
-
-export type ParamFieldDefinition = Exclude<
-  FieldDefinition,
-  { type: 'boolean' } | { type: 'object' } | { type: 'array' }
->
-
-export type QueryFieldDefinition = Exclude<FieldDefinition, { type: 'object' } | { type: 'array' }>
-
-export interface RouteDefinition {
-  method: 'get' | 'post' | 'put' | 'delete' | 'patch'
-  endpoint: string
-  description?: string
-  tags?: string[]
-  params?: Record<string, ParamFieldDefinition>
-  query?: Record<string, QueryFieldDefinition>
-  body?: Record<string, FieldDefinition>
-  responses?: Array<{
-    status: number
-    description: string
-    schema?: Record<string, FieldDefinition>
-  }>
-  middlewares?: MiddlewareDescriptor[]
-  scope?: 'acl'
+export type RouteDefinition = Omit<BaseRouteDefinition<MiddlewareDescriptor>, 'callback'> & {
   callback: (params: {
     req: Request
     res: Response
@@ -89,5 +36,4 @@ export interface RouteDefinition {
     restApi: RestApi
     createAction: CreateAction
   }) => MaybePromise<unknown>
-  security?: SecurityName
 }
